refactor(esp32): Simplify and reorganize component managers
Refactorización de componentes existentes:

- **EthernetManager**:
  - Movido de components/network/ a main/
  - Simplificado: eliminadas 97 líneas de código redundante
  - Mejor integración con hardware_config.h
  - Optimizado para uso con W5500

- **CANManager**:
  - Simplificado: reducidas 100+ líneas
  - Eliminada lógica compleja innecesaria
  - Enfoque en funcionalidad core CiA402
  - Mejor manejo de errores

- **WiFiManager**:
  - Ajustes menores de compatibilidad
  - Actualización de logs

- **ConfigManager**:
  - Correcciones menores
  - Mejor validación de parámetros

Esta refactorización unifica la arquitectura y reduce complejidad innecesaria,
preparando el sistema para la arquitectura multitarea FreeRTOS.

Deleted: esp32_gateway/components/network/* (código movido a main/)
Modified: 4 managers simplificados (-200 líneas aproximadamente)

+++ esp32_gateway/main/config_manager.cpp
@@ -44,8 +44,8 @@ esp_err_t ConfigManager::load_can_config(CANConfig* config) {
     }
 
     nvs_close(handle);
-    ESP_LOGI(TAG, "Configuración CAN cargada: NodeID=0x%02X, Bitrate=%d kbps", 
-             config->node_id, config->bitrate);
+    ESP_LOGI(TAG, "Configuración CAN cargada: NodeID=0x%02X, Bitrate=%lu kbps",
+             config->node_id, (unsigned long)config->bitrate);
     return ESP_OK;
 }
